fixed some models

+++ SpaceTraders_SDK/src/main/java/io/github/yamin8000/spacetraders_sdk/model/ship/Condition.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.Serializable
 import kotlin.jvm.JvmInline
 
 /**
- * Condition is a range of 0 to 100 where 0 is completely worn out and 100 is brand new.
+ * Condition is a range of 0 to 1 where 0 is completely worn out and 1 is brand new.
  */
 @Serializable
 @JvmInline
-value class Condition(val value: Int)
+value class Condition(val value: Float)
